structure rename for readability
